energy/victron-multiplus-ii/pyproject.toml
[tool.poetry]
name = "victron-multiplus-ii"
version = "0.1.0"
description = ""
authors = ["Andrew LeCody <andrew@beyondweb.net>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
pyserial = "^3.5"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
